Rewriting C# scripting v3.8 (WIP)
Made it so that it loads the value that the C# has and can change the value now, made it so that more types can be registered too

Bug/Intended feature? : When there's multiple scripts, there will be a conflict, eg. automoveX and playertest, the player will not move. I don't think this will be a problem

To Do : The change in variable only changes when the play button is played which idt that should be the case

To Do : Serialization of the script

To Do : Move the functions around so it looks neater away from ImGuiSceneHierarchy